fix size error in tests

+++ test/exe/main_common_test.cc
@@ -109,7 +109,7 @@ TEST_P(MainCommonTest, ConstructDestructHotRestartDisabledNoInitWithVectorArgs) 
   addArg("--disable-hot-restart");
   initOnly();
   std::vector<std::string> args(argv_.size());
-  for (int i = 0; i < argv_.size() - 1; ++i) {
+  for (size_t i = 0; i < argv_.size() - 1; ++i) {
     args[i] = std::string(argv_[i]);
   }
   MainCommon main_common(args);
